Remove new StatusChangeEvent - link to current PresenceUpdateEvent
